pyproject.toml
[project]
name = "lawqa-rag-studio"
version = "0.1.0"
description = "LawQA-RAG-Studio benchmarking and serving framework"
authors = [{ name = "LawQA-RAG-Studio Team" }]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6",
    "typer>=0.12",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "openai>=1.57",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "numpy>=1.26",
    "qdrant-client>=1.12",
    "fastembed>=0.7.4",
    "transformers>=4.41",
    "torch>=2.2",
    "fugashi[unidic-lite]>=1.3",
]

[project.scripts]
lawqa-rag-studio = "lawqa_rag_studio.cli:main"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
